Placeholder for picker, Homescreen, Navigation
Created a custom placeholder for the picker component. This placeholder is suppose to be temporary as it has a flaw. A new homescreen is created which is
empty for now. It is created to test the react-navigation library which is also installed and currently being tested. Currently, everything works as it
should.