Sectionlist Implementation In HomeScreen
Created a file called expensesTable which queries the added expenses from expenses table in GastoCalc.db. Then in the homeScreen file, I formate this data
make it useable in a Sectionlist as per needs. This Sectionlist is now temporarily being printed to check the working of the programe. Further formating
will be done in the next commit.