Bug Fixes + editExpense Screen
editExpense screen is now complete. It takes the data and can either edit or delete it from the expenses table in the database. Further, now the
dailyExpense gets updated evertime a new expense is added or a change has been made in an existing expense. This is acheived by calculated
dailyExpense individually rather than creating an array initially and then printing it later.